fixed an error when deleting merge.log

+++ Project01/merge_log.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-rm merge.log
+if [[ -f merge.log ]]; then
+  rm merge.log
+fi
 git_log=$(git log --oneline | grep -i "merge" | cut -d' ' -f1)
 echo "$git_log" >> merge.log
